chore: tidy nav bar refactor: abstract big components into their own
