+++ src/components/ui/colors.ts
@@ -1,8 +1,8 @@
 export const customColors = {
   brand: {
     main: '#7c70d2', // purple
-    main400: '#6658cb', // darker purple
-    main800: '#867bd5', // lighter purple
+    main400: '#867bd5', // lighter purple
+    main800: '#6359b8', // darker purple
     dark: '#1c3177', //dark blue
     light: '#ddd8fe', //light purple
     dark800: '#101d47', //darker blue
